strategy returns void, ioctl cmd is u_long
SCCS-vsn: sys/dev/vn.c 8.8
SCCS-vsn: sys/dev/cd.c 8.3
SCCS-vsn: sys/hp/dev/grf.c 8.5
SCCS-vsn: sys/hp/dev/dca.c 8.3
SCCS-vsn: sys/hp/dev/hil.c 8.3
SCCS-vsn: sys/hp/dev/ite.c 8.3
SCCS-vsn: sys/hp/dev/cons.c 8.3

+++ usr/src/sys/hp/dev/ite.c
@@ -11,7 +11,7 @@
  *
  * from: Utah $Hdr: ite.c 1.28 92/12/20$
  *
- *	@(#)ite.c	8.2 (Berkeley) %G%
+ *	@(#)ite.c	8.3 (Berkeley) %G%
  */
 
 /*
@@ -261,7 +261,7 @@ itewrite(dev, uio, flag)
 
 iteioctl(dev, cmd, addr, flag, p)
 	dev_t dev;
-	int cmd;
+	u_long cmd;
 	caddr_t addr;
 	int flag;
 	struct proc *p;
